agent/service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.2.16
langchain-openai==0.1.23
python-dotenv==1.0.1
pytest==8.3.3
socksio>=1.0.0
